requirements.txt
requests
PyYAML
twitchio
py-logs